Modify voice recognition algorithm : DTW --> correlation

+++ Application-images/voice_recognition_sim/src/dtw.h
@@ -3,8 +3,6 @@
 
 #include <stdint.h>
 
-// Compare 2 signaux audio et retourne la "distance"
-// Plus petite distance = plus similaire
-float dtw_distance(const uint8_t *signal1, const uint8_t *signal2, int length);
+uint32_t dtw_distance(const uint8_t *signal1, const uint8_t *signal2, int length);
 
 #endif
